chanages

+++ src/models/orders.js
@@ -87,13 +87,13 @@ module.exports = (sequelize, Sequelize) => {
           type:Sequelize.DATE
         },
         IGST:{
-          type: Sequelize.INTEGER
+          type: Sequelize.DOUBLE
         },
         CGST:{
-          type: Sequelize.INTEGER
+          type: Sequelize.DOUBLE
         },
         SGST:{
-          type: Sequelize.INTEGER
+          type: Sequelize.DOUBLE
         },
         subTotal:{
           type: Sequelize.STRING
